refactor: expose a function to get connect method order with appkit instance, refactor public state controller
